fix(benchmarks): isolate Jackson 3 to Java 17
